+++ noodles-gff/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.41.0 - 2024-12-20
 
 ### Changed